Move user feedback and tag models into Admin.Common
