+++ Apps/Admin/Server/Services/UserFeedbackService.cs
@@ -17,8 +17,8 @@ namespace HealthGateway.Admin.Server.Services
 {
     using System;
     using System.Collections.Generic;
+    using HealthGateway.Admin.Common.Models;
     using HealthGateway.Admin.Server.Converters;
-    using HealthGateway.Admin.Server.Models;
     using HealthGateway.Common.Data.Constants;
     using HealthGateway.Common.Data.ViewModels;
     using HealthGateway.Database.Constants;
